Change to use global signal
