+++ src/pretix/control/navigation.py
@@ -569,11 +569,6 @@ def get_admin_navigation(request):
                     'label': _('Generate keys for SSO'),
                     'url': reverse('control:admin.global.sso'),
                     'active': (url.url_name == 'admin.global.sso'),
-                },
-                {
-                    'label': _('Social login settings'),
-                    'url': reverse('control:admin.global.social.auth.settings'),
-                    'active': (url.url_name == 'admin.global.social.auth.settings'),
                 }
             ]
         }
