+++ src/pretix/plugins/socialauth/apps.py
@@ -14,6 +14,10 @@ class SocialAuthApp(AppConfig):
         version = version
         featured = True
         description = _("This plugin allows you to login via social networks")
+        visible = False
+
+    def ready(self):
+        from . import signals  # NOQA
 
 
-default_app_config = 'pretix.plugins.socialauth.PaypalApp'
+default_app_config = 'pretix.plugins.socialauth.SocialAuthApp'
