+++ src/pretix/plugins/socialauth/templates/socialauth/social_auth_settings.html
