+++ src/pretix/plugins/socialauth/views.py
@@ -5,8 +5,12 @@ from allauth.socialaccount.adapter import get_adapter
 from django.conf import settings
 from django.contrib import messages
 from django.shortcuts import redirect
+from django.urls import reverse
+from django.views.generic import TemplateView
 
 from pretix.base.models import User
+from pretix.base.settings import GlobalSettingsObject
+from pretix.control.permissions import AdministratorPermissionRequiredMixin
 from pretix.control.views.auth import process_login_and_set_cookie
 from pretix.helpers.urls import build_absolute_uri
 
@@ -42,3 +46,37 @@ def oauth_return(request):
         )
         logger.error('Error while authorizing: user has no email address.')
         return redirect('control:auth.login')
+
+
+class SocialLoginView(AdministratorPermissionRequiredMixin, TemplateView):
+    template_name = 'socialauth/social_auth_settings.html'
+    LOGIN_PROVIDERS = {'mediawiki': False, 'github': False, 'google': False}
+    VALID_STATES = {'enable', 'disable'}
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.gs = GlobalSettingsObject()
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        if self.gs.settings.get('login_providers', as_type=dict) is None:
+            self.gs.settings.set('login_providers', self.LOGIN_PROVIDERS)
+        context['login_providers'] = self.gs.settings.get('login_providers', as_type=dict)
+        return context
+
+    def post(self, request, *args, **kwargs):
+        login_providers = self.gs.settings.get('login_providers', as_type=dict)
+        for provider in self.LOGIN_PROVIDERS.keys():
+            value = request.POST.get(f'{provider}_login', '').lower()
+            if value not in self.VALID_STATES:
+                continue
+            elif value == "enable":
+                login_providers[provider] = True
+            elif value == "disable":
+                login_providers[provider] = False
+            self.gs.settings.set('login_providers', login_providers)
+            break
+        return redirect(self.get_success_url())
+
+    def get_success_url(self) -> str:
+        return reverse('plugins:socialauth:admin.global.social.auth.settings')
